Extract BrowserSwitchLauncher.
